Separating history object from reducer
When you use this example to put it in a real world redux app you might want to split reducer and router into separate files. I think the best way to deal with the history object is leave it out of the reducer and check at the render method if `location` and `action` is undefined.

Another option would be to put the history object in the store as well (does not feel very reduxish) or to instantiate it twice (yikes).